test(client): fix running e2e tests in parallel
When running tests locally, unless the `--maxWorkers` option was passed,
all tests were started at the same time. This caused most of them to
fail because Docker ran out of IP address ranges to assign to the
networks created by Docker Compose.

This commit changes the default value of `--maxWorkers` from `Infinity`
to `os.cpus().length`, and moves the hardcoded alternative default that
was applied when `CI=true` environment variable was set to the CLI
command used in the CI workflow as an explicit option.

Additionally, the pooling implementation was inefficient, it was
unnecessary looping every 100 ms to check if the next iteration can
start, instead of starting as soon as it becomes possible and not doing
anything before that. I refactored it to a proper semaphore.